+++ src/ninja/mbedded/ninjaterm/interfaces/RawDataReceivedListener.java
@@ -1,8 +1,0 @@
-package ninja.mbedded.ninjaterm.interfaces;
-
-/**
- * Created by gbmhu on 2016-09-22.
- */
-public interface RawDataReceivedListener {
-    void update(String data);
-}