+++ changelog.md
@@ -7,6 +7,7 @@ vX.X.X
 - Fixed bug where ANSI escape codes were not disabled when it's checkbox is deselected, closes #119.
 - Added bits/second metrics to the "Stats" sub-tab, closes #120.
 - Added bottom bar to the status pane which displays the total TX/RX bytes per second rates, closes #122.
+- Moved interface files from "interfaces/" and into the package/folder that they have most relevance with, closes #121.
 
 v0.6.2
 ------